requirements.txt
streamlit
google-generativeai
python-dotenv
vertexai
sqlite3
pandas